checkpoint, failed clickwrite
