+++ src/config/mod.rs
@@ -7,6 +7,8 @@ pub use self::click::ClickConfig;
 pub use self::click::CompletionType;
 pub use self::click::EditMode;
 
+#[cfg(test)]
+pub use self::kube::tests::get_test_config;
 pub use self::kube::Config;
 
 pub use self::kubefile::AuthProvider;
